feat: extract permissions pull list component and add tests
